PCHR-2813: Move drupal mocks to an include file and remove unnecessary includes
